@@ -30,7 +30,7 @@ int main() {
 
 void prepare_data() {
     string file_name = "train_data.txt";
-    dbm::make_data<float>(file_name, 100000, 30, 'b');
+    dbm::make_data<float>(file_name, 100000, 30, 't');
 }
 
 void test_save_load_dbm() {
@@ -58,8 +58,8 @@ void test_save_load_dbm() {
     dbm::Matrix<float> smaller_prediction(int(0.75 * n_samples), 1, 0);
 
     // ================
-    string param_string = "no_learners 100 no_candidate_feature 5 loss_function b "
-            "no_train_sample 10000 max_depth 5 no_candidate_split_point 5";
+    string param_string = "no_learners 500 no_candidate_feature 5 loss_function t "
+            "no_train_sample 50000 max_depth 5 no_candidate_split_point 5";
     dbm::DBM<float> dbm(param_string);
 
     {
@@ -97,55 +97,6 @@ void test_save_load_dbm() {
     dbm::Matrix<float> combined = dbm::hori_merge(prediction, re_prediction);
     dbm::Matrix<float> result = dbm::hori_merge(train_y, combined);
     result.print_to_file("whole_result.txt");
-}
-
-void test_save_load_tree() {
-    int n_samples = 100000, n_features = 30, n_width = 31;
-
-    dbm::Matrix<float> train_data(n_samples, n_width, "train_data.txt");
-    dbm::Matrix<float> prediction(n_samples, 1, 0);
-
-    int row_inds[n_samples], col_inds[n_features];
-
-    for (int i = 0; i < n_features; ++i)
-        col_inds[i] = i;
-    for (int i = 0; i < n_samples; ++i)
-        row_inds[i] = i;
-
-    dbm::Matrix<float> train_x = train_data.cols(col_inds, n_features);
-    dbm::Matrix<float> train_y = train_data.col(n_features);
-
-    // ========================================================
-
-    dbm::Params params = dbm::set_params("max_depth 4 no_candidate_split_point 5");
-
-    dbm::Tree_node<float> *tree = new dbm::Tree_node<float>(0);
-    dbm::Tree_trainer<float> trainer(params);
-
-    {
-        dbm::Time_measurer time_measurer;
-        trainer.train(tree, train_x, train_y, prediction, nullptr, 'n', row_inds, n_samples, col_inds, n_features);
-//        trainer.prune(tree);
-    }
-
-    {
-        dbm::Tree_info<float> info(tree);
-        info.print_to_file("tree.txt", 0);
-    }
-
-    {
-        ofstream out("save.txt");
-        dbm::save_tree_node(tree, out);
-    }
-
-
-    dbm::Tree_node<float> *re_tree;
-    {
-        ifstream in("save.txt");
-        dbm::load_tree_node(in, re_tree);
-        dbm::Tree_info<float> info(re_tree);
-        info.print_to_file("re_tree.txt", 0);
-    }
 }
 
 void train_a_dbm() {
@@ -197,12 +148,67 @@ void train_a_dbm() {
     result.print_to_file("whole_result.txt");
 }
 
+void test_save_load_tree() {
+    int n_samples = 100000, n_features = 30, n_width = 31;
+
+    dbm::Matrix<float> train_data(n_samples, n_width, "train_data.txt");
+    dbm::Matrix<float> prediction(n_samples, 1, 0);
+    dbm::Matrix<float> ind_delta(n_samples, 1, 0);
+
+    int row_inds[n_samples], col_inds[n_features];
+
+    for (int i = 0; i < n_features; ++i)
+        col_inds[i] = i;
+    for (int i = 0; i < n_samples; ++i)
+        row_inds[i] = i;
+
+    dbm::Matrix<float> train_x = train_data.cols(col_inds, n_features);
+    dbm::Matrix<float> train_y = train_data.col(n_features);
+
+    // ========================================================
+
+    dbm::Params params = dbm::set_params("max_depth 4 no_candidate_split_point 5");
+
+    dbm::Tree_node<float> *tree = new dbm::Tree_node<float>(0);
+    dbm::Tree_trainer<float> trainer(params);
+
+    dbm::Loss_function<float> loss_function(params);
+    loss_function.calculate_ind_delta(train_y, prediction,
+                                      ind_delta, params.loss_function, row_inds, n_samples);
+
+    {
+        dbm::Time_measurer time_measurer;
+        trainer.train(tree, train_x, train_y, ind_delta, prediction, nullptr, 'n', row_inds, n_samples, col_inds, n_features);
+//        trainer.prune(tree);
+    }
+
+    {
+        dbm::Tree_info<float> info(tree);
+        info.print_to_file("tree.txt", 0);
+    }
+
+    {
+        ofstream out("save.txt");
+        dbm::save_tree_node(tree, out);
+    }
+
+
+    dbm::Tree_node<float> *re_tree;
+    {
+        ifstream in("save.txt");
+        dbm::load_tree_node(in, re_tree);
+        dbm::Tree_info<float> info(re_tree);
+        info.print_to_file("re_tree.txt", 0);
+    }
+}
+
 void build_a_tree() {
 
     int n_samples = 100000, n_features = 30, n_width = 31;
 
     dbm::Matrix<float> train_data(n_samples, n_width, "train_data.txt");
     dbm::Matrix<float> prediction(n_samples, 1, 0);
+    dbm::Matrix<float> ind_delta(n_samples, 1, 0);
 
     int row_inds[n_samples], col_inds[n_features];
 
@@ -223,7 +229,11 @@ void build_a_tree() {
 
     dbm::Time_measurer time_measurer;
 
-    trainer.train(tree, train_x, train_y, prediction, nullptr, 'n', row_inds, n_samples, col_inds, n_features);
+    dbm::Loss_function<float> loss_function(params);
+    loss_function.calculate_ind_delta(train_y, prediction,
+                                      ind_delta, params.loss_function, row_inds, n_samples);
+
+    trainer.train(tree, train_x, train_y, ind_delta, prediction, nullptr, 'n', row_inds, n_samples, col_inds, n_features);
 
     time_measurer.~Time_measurer();
 
